Simulated Data for TcGSA

Simulated data for 5 gene sets of 50 genes each. Gene expression is simulated at 5 time points for 10 patients.

Usage

data(data_simu_TcGSA)

Value

expr_1grp

See Details.

expr_2grp

See Details.

design

a data frame with 5 variables:

Patient_ID: a factor that contains the patient ID.

TimePoint: a numeric vector or a factor that contains the time points at which gene expression was measured.

sample_name: a character vector with the names of the sample (corresponding to the names of the columns of expr_1grp and of expr_2grp).

group.var: a factor that indicates to which of the 2 treatment groups each sample belongs to.

Group_paired_ID a random paired identifier for paired couples (one in each of the 2 treatment groups) of patients.

gmt_sim

a gmt object containing the gene sets definition. See GSA.read.gmt and GMT definition on www.broadinstitute.org.

Details

In expr_1grp all patients belong to the same unique treatment group. The first 2 gene sets are simulated under the null hypothesis. The gene sets 3 and 4 are simulated under the alternative hypothesis that there is a significant homogeneous time trend within the gene set. The gene set 5 is simulated under the alternative hypothesis that there are significant heterogeneous time trends within the gene set.

In expr_2grp all patients belong to 2 treatment groups. The 5 first patients belong to the treatment group 'T', The 5 other patients belong to the treatment group 'C'. The first 2 gene sets are simulated under the null hypothesis that there is no difference in the time trend between the 2 treatment groups. The gene sets 3 and 4 are simulated under the alternative hypothesis that there are significantly different homogeneous time trends within the gene set between the 2 treatment groups. The gene set 5 is simulated under the alternative hypothesis that there are significantly different heterogeneous time trends between the 2 treatment groups within the gene set.

Source

This is simulated data.

Author

Boris P. Hejblum

See also

Examples

data(data_simu_TcGSA)
summary(expr_1grp)
#>        1                2                3                4         
#>  Min.   : 3.205   Min.   : 3.309   Min.   : 3.367   Min.   : 3.330  
#>  1st Qu.: 5.251   1st Qu.: 5.299   1st Qu.: 5.184   1st Qu.: 5.090  
#>  Median : 6.256   Median : 6.300   Median : 6.172   Median : 6.156  
#>  Mean   : 6.420   Mean   : 6.481   Mean   : 6.413   Mean   : 6.407  
#>  3rd Qu.: 7.524   3rd Qu.: 7.600   3rd Qu.: 7.604   3rd Qu.: 7.674  
#>  Max.   :11.231   Max.   :10.798   Max.   :10.278   Max.   :11.288  
#>        5                6                7                8         
#>  Min.   : 3.234   Min.   : 3.363   Min.   : 3.399   Min.   : 3.427  
#>  1st Qu.: 4.977   1st Qu.: 5.233   1st Qu.: 5.270   1st Qu.: 5.201  
#>  Median : 6.048   Median : 6.279   Median : 6.247   Median : 6.149  
#>  Mean   : 6.400   Mean   : 6.483   Mean   : 6.442   Mean   : 6.468  
#>  3rd Qu.: 7.927   3rd Qu.: 7.734   3rd Qu.: 7.689   3rd Qu.: 7.761  
#>  Max.   :10.522   Max.   :10.278   Max.   :10.181   Max.   :10.172  
#>        9                10               11               12        
#>  Min.   : 3.395   Min.   : 3.172   Min.   : 3.395   Min.   : 3.442  
#>  1st Qu.: 5.177   1st Qu.: 5.033   1st Qu.: 5.193   1st Qu.: 5.198  
#>  Median : 6.121   Median : 5.998   Median : 6.218   Median : 6.246  
#>  Mean   : 6.458   Mean   : 6.455   Mean   : 6.368   Mean   : 6.406  
#>  3rd Qu.: 7.800   3rd Qu.: 7.858   3rd Qu.: 7.610   3rd Qu.: 7.705  
#>  Max.   :10.271   Max.   :10.351   Max.   :10.004   Max.   :10.384  
#>        13               14               15               16        
#>  Min.   : 3.415   Min.   : 3.416   Min.   : 3.251   Min.   : 3.349  
#>  1st Qu.: 5.162   1st Qu.: 5.021   1st Qu.: 4.867   1st Qu.: 5.401  
#>  Median : 6.094   Median : 6.102   Median : 5.923   Median : 6.287  
#>  Mean   : 6.373   Mean   : 6.380   Mean   : 6.357   Mean   : 6.468  
#>  3rd Qu.: 7.667   3rd Qu.: 7.707   3rd Qu.: 7.802   3rd Qu.: 7.617  
#>  Max.   :11.174   Max.   :10.410   Max.   :10.164   Max.   :10.382  
#>        17               18               19               20        
#>  Min.   : 3.375   Min.   : 2.718   Min.   : 3.398   Min.   : 3.334  
#>  1st Qu.: 5.348   1st Qu.: 5.275   1st Qu.: 5.187   1st Qu.: 5.168  
#>  Median : 6.310   Median : 6.253   Median : 6.227   Median : 6.171  
#>  Mean   : 6.438   Mean   : 6.430   Mean   : 6.413   Mean   : 6.434  
#>  3rd Qu.: 7.617   3rd Qu.: 7.609   3rd Qu.: 7.705   3rd Qu.: 7.755  
#>  Max.   :10.239   Max.   :10.525   Max.   :10.353   Max.   :10.352  
#>        21               22               23               24        
#>  Min.   : 3.210   Min.   : 3.281   Min.   : 3.260   Min.   : 3.319  
#>  1st Qu.: 5.263   1st Qu.: 5.274   1st Qu.: 5.151   1st Qu.: 5.096  
#>  Median : 6.347   Median : 6.321   Median : 6.239   Median : 6.206  
#>  Mean   : 6.498   Mean   : 6.518   Mean   : 6.481   Mean   : 6.480  
#>  3rd Qu.: 7.884   3rd Qu.: 7.891   3rd Qu.: 7.708   3rd Qu.: 7.939  
#>  Max.   :10.680   Max.   :10.096   Max.   :11.141   Max.   :10.521  
#>        25               26               27               28        
#>  Min.   : 3.189   Min.   : 3.317   Min.   : 3.372   Min.   : 3.359  
#>  1st Qu.: 5.051   1st Qu.: 5.362   1st Qu.: 5.279   1st Qu.: 5.269  
#>  Median : 6.078   Median : 6.388   Median : 6.294   Median : 6.240  
#>  Mean   : 6.480   Mean   : 6.476   Mean   : 6.447   Mean   : 6.465  
#>  3rd Qu.: 7.916   3rd Qu.: 7.702   3rd Qu.: 7.682   3rd Qu.: 7.721  
#>  Max.   :10.652   Max.   :10.468   Max.   :10.251   Max.   :10.418  
#>        29               30               31               32        
#>  Min.   : 3.429   Min.   : 3.287   Min.   : 3.339   Min.   : 3.350  
#>  1st Qu.: 5.197   1st Qu.: 5.055   1st Qu.: 5.223   1st Qu.: 5.212  
#>  Median : 6.136   Median : 6.078   Median : 6.158   Median : 6.258  
#>  Mean   : 6.442   Mean   : 6.440   Mean   : 6.364   Mean   : 6.388  
#>  3rd Qu.: 7.759   3rd Qu.: 7.826   3rd Qu.: 7.456   3rd Qu.: 7.498  
#>  Max.   :10.677   Max.   :11.181   Max.   :10.621   Max.   :10.629  
#>        33               34               35               36        
#>  Min.   : 3.412   Min.   : 3.371   Min.   : 3.219   Min.   : 3.243  
#>  1st Qu.: 5.181   1st Qu.: 5.007   1st Qu.: 5.070   1st Qu.: 5.306  
#>  Median : 6.171   Median : 6.093   Median : 5.971   Median : 6.283  
#>  Mean   : 6.381   Mean   : 6.335   Mean   : 6.360   Mean   : 6.441  
#>  3rd Qu.: 7.606   3rd Qu.: 7.603   3rd Qu.: 7.612   3rd Qu.: 7.579  
#>  Max.   :10.165   Max.   :10.578   Max.   :10.541   Max.   :10.652  
#>        37              38               39               40        
#>  Min.   :3.354   Min.   : 3.299   Min.   : 3.321   Min.   : 3.269  
#>  1st Qu.:5.313   1st Qu.: 5.223   1st Qu.: 5.158   1st Qu.: 5.138  
#>  Median :6.191   Median : 6.219   Median : 6.185   Median : 6.067  
#>  Mean   :6.432   Mean   : 6.449   Mean   : 6.459   Mean   : 6.456  
#>  3rd Qu.:7.684   3rd Qu.: 7.678   3rd Qu.: 7.720   3rd Qu.: 7.751  
#>  Max.   :9.953   Max.   :10.027   Max.   :10.336   Max.   :10.254  
#>        41               42               43               44        
#>  Min.   : 3.125   Min.   : 3.188   Min.   : 3.235   Min.   : 3.218  
#>  1st Qu.: 5.149   1st Qu.: 5.192   1st Qu.: 5.143   1st Qu.: 5.098  
#>  Median : 6.191   Median : 6.185   Median : 6.113   Median : 6.052  
#>  Mean   : 6.381   Mean   : 6.407   Mean   : 6.377   Mean   : 6.401  
#>  3rd Qu.: 7.664   3rd Qu.: 7.656   3rd Qu.: 7.758   3rd Qu.: 7.809  
#>  Max.   :10.375   Max.   :10.268   Max.   :10.265   Max.   :10.283  
#>        45               46               47               48        
#>  Min.   : 3.223   Min.   : 3.190   Min.   : 3.264   Min.   : 3.237  
#>  1st Qu.: 4.916   1st Qu.: 5.212   1st Qu.: 5.264   1st Qu.: 5.102  
#>  Median : 6.037   Median : 6.260   Median : 6.234   Median : 6.174  
#>  Mean   : 6.363   Mean   : 6.378   Mean   : 6.391   Mean   : 6.404  
#>  3rd Qu.: 7.856   3rd Qu.: 7.489   3rd Qu.: 7.482   3rd Qu.: 7.660  
#>  Max.   :10.220   Max.   :10.504   Max.   :10.623   Max.   :10.459  
#>        49               50        
#>  Min.   : 3.252   Min.   : 3.247  
#>  1st Qu.: 5.080   1st Qu.: 5.075  
#>  Median : 6.113   Median : 6.002  
#>  Mean   : 6.397   Mean   : 6.364  
#>  3rd Qu.: 7.733   3rd Qu.: 7.839  
#>  Max.   :10.596   Max.   :10.508  
summary(design)
#>    Patient_ID   TimePoint  Sample_name group.var Group_ID_paired
#>  P1     : 5   Min.   :1   1      : 1   C:25      cpl1:10        
#>  P10    : 5   1st Qu.:2   10     : 1   T:25      cpl2:10        
#>  P2     : 5   Median :3   11     : 1             cpl3:10        
#>  P3     : 5   Mean   :3   12     : 1             cpl4:10        
#>  P4     : 5   3rd Qu.:4   13     : 1             cpl5:10        
#>  P5     : 5   Max.   :5   14     : 1                            
#>  (Other):20               (Other):44                            
gmt_sim
#> $genesets
#> $genesets[[1]]
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> 
#> $genesets[[2]]
#>  [1]  51  52  53  54  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69
#> [20]  70  71  72  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88
#> [39]  89  90  91  92  93  94  95  96  97  98  99 100
#> 
#> $genesets[[3]]
#>  [1] 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119
#> [20] 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138
#> [39] 139 140 141 142 143 144 145 146 147 148 149 150
#> 
#> $genesets[[4]]
#>  [1] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168 169
#> [20] 170 171 172 173 174 175 176 177 178 179 180 181 182 183 184 185 186 187 188
#> [39] 189 190 191 192 193 194 195 196 197 198 199 200
#> 
#> $genesets[[5]]
#>  [1] 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219
#> [20] 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238
#> [39] 239 240 241 242 243 244 245 246 247 248 249 250
#> 
#> 
#> $geneset.names
#> [1] "Gene set 1" "Gene set 2" "Gene set 3" "Gene set 4" "Gene set 5"
#> 
#> $geneset.descriptions
#> [1] "This is a simulated gene set under H0"                
#> [2] "This is a simulated gene set under H0"                
#> [3] "This is a simulated gene set under H1 (homogeneous)"  
#> [4] "This is a simulated gene set under H1 (homogeneous)"  
#> [5] "This is a simulated gene set under H1 (heterogeneous)"
#> 
#> attr(,"class")
#> [1] "GSA.genesets"