Summarizing TcGSA

summary method for class 'TcGSA'

Usage

# S3 method for class 'TcGSA'
summary(object, ...)

# S3 method for class 'summary.TcGSA'
print(x, ...)

Arguments

object

an object of class 'TcGSA'.

...

further arguments passed to or from other methods.

x

an object of class 'summary.TcGSA'.

Value

The function summary.TcGSA returns a list with the following components (list elements):

time_func the chosen form for the time trend.

separateSubjects a logical flag indicating whether gene sets tested for discriminating among patients, or for time trends over time.

ntg the number of treatment groups.

ngs the number of tested gene sets.

nsignif the number of significant gene sets at a 5% FDR (using the default Benjamini & Yekutieli step-up procedure).

See also

Author

Boris P. Hejblum

Examples

if(interactive()){
data(data_simu_TcGSA)

tcgsa_sim_1grp <- TcGSA.LR(expr=expr_1grp, gmt=gmt_sim, design=design, 
                          subject_name="Patient_ID", time_name="TimePoint",
                          time_func="linear", crossedRandom=FALSE)
summary(tcgsa_sim_1grp)

tcgsa_sim_2grp <- TcGSA.LR(expr=expr_2grp, gmt=gmt_sim, design=design, 
                          subject_name="Patient_ID", time_name="TimePoint",
                          time_func="linear", crossedRandom=FALSE, 
                          group_name="group.var")
summary(tcgsa_sim_2grp)
}